chore: relocate tests/unit mirrors of modules moved in Phases 3-4
The tests/unit tree mirrors the source tree, so specs must follow their
sources: 15 specs orphaned in stale mirrors (12 from Phase 4, 3 from
Phase 3's REST API move) now live under tests/unit/server/** mirroring
the new source locations, with proxyquire targets recomputed. No glob
changes needed: the existing tests/unit/server/** globs already cover
the new locations, and the tests/unit/app/** globs still match the
remaining specs.

Adds a "Mirrored Test Trees" section to MIGRATION_PLAN.md so the
remaining phases move mirrored specs together with their sources.
