+++ apps/meteor/tests/unit/server/lib/users/saveUser/sendUserEmail.spec.ts
@@ -54,7 +54,7 @@ describe('sendUserEmail (Mocha + TS)', () => {
 		SettingsStub = stubs.SettingsStub as any;
 		MeteorStub = stubs.MeteorStub as any;
 
-		const mod = mock.noCallThru().load('./../../../../../../../meteor/server/lib/users/saveUser/sendUserEmail.ts', {
+		const mod = mock.noCallThru().load('../../../../../../server/lib/users/saveUser/sendUserEmail.ts', {
 			'../../../../app/mailer/server/api': MailerStub,
 			'../../../../app/settings/server': SettingsStub,
 			'meteor/meteor': MeteorStub,
